split webconnectivitylte tests
